% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{complete}
\alias{complete}
\title{Extract completed imputed data}
\usage{
complete(x, ...)
}
\arguments{
\item{x}{An object containing completed data.}

\item{...}{Passed to methods.}
}
\value{
A data frame or list of data frames.
}
\description{
\code{complete()} extracts completed data sets from imputation objects. For
\code{mimar_imputation} objects, use \code{complete(x)} or \code{complete(x, 1)} for one
completed data set, \code{complete(x, "all")} for all completed data sets, and
\code{complete(x, "long")} for a stacked long data frame.
}
